+++ utils/TaxaFuncPloter/bar_plot_js.py
@@ -74,7 +74,7 @@ class BarPlot_js:
             import seaborn as sns
             colors = sns.color_palette('deep', col_num)
             colors = [f'rgb({int(i[0]*255)},{int(i[1]*255)},{int(i[2]*255)})' for i in colors]
-        
+               
         # create title
         if title is None:
             if taxon_name is None:
@@ -96,10 +96,8 @@ class BarPlot_js:
             color = colors[i]
             c.add_yaxis(name, list(df.loc[name, :]), 
                         stack="stack1", 
-                        # category_gap="50%",
-                        category_gap="5%",
-                        # itemstyle_opts=opts.ItemStyleOpts(color=color, border_color="black", border_width=0.1)
-                        itemstyle_opts=opts.ItemStyleOpts(color=color, border_type=None)
+                        category_gap="5%", 
+                        itemstyle_opts=opts.ItemStyleOpts(color=color, border_type=None, border_width=0)
                         )
 
 
@@ -107,6 +105,7 @@ class BarPlot_js:
 
         params = {
             "legend_opts": opts.LegendOpts(
+                type_="scroll", page_icon_size = 8, selector=[{"type": "all", "title": "All"}, {"type": "inverse", "title": "Inverse"}],
                 pos_left="right", orient="vertical", pos_top="5%", border_width=0
             )
             if show_legend
@@ -119,11 +118,15 @@ class BarPlot_js:
                 )
             ],
             "toolbox_opts": opts.ToolboxOpts(
-                is_show=True, orient="vertical", pos_left="right", pos_top="bottom"
+                is_show=True, orient="vertical", pos_left="left", pos_top="bottom"
             ),
             "title_opts": opts.TitleOpts(title=f"{title}", pos_left="center"),
             "xaxis_opts": opts.AxisOpts(
                 axislabel_opts=opts.LabelOpts(rotate=25, font_size=font_size)
+            ),
+            "yaxis_opts": opts.AxisOpts(
+                axislabel_opts=opts.LabelOpts(font_size=font_size),
+                max_=100.02 if plot_percent else None, # set max value to 100.02 rather than 100: the sum of some columns may more than 100 due to float number
             ),
         }
         c.set_global_opts(**params)